Break up test_distributions.py
