+++ pymc/tests/distributions/test_dist_math.py
