+++ pymc/tests/distributions/test_moments.py
