+++ pymc/tests/distributions/test_shape_utils.py
